import Batch fragment contextual tool bar functionality like update/create/delete

+++ app/src/main/java/com/example/myittaroostockinventorymanger/local/Stock.kt
@@ -13,5 +13,5 @@ data class Stock(
 ) : Parcelable {
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "stock_id")
-     var stockId: Int =0
+    var stockId: Long = 0
 }
